# Optimize for SSSE3