--- a/src/fx.rs
+++ b/src/fx.rs
@@ -26,14 +26,14 @@ pub struct FxHasher {
 }
 
 #[cfg(target_pointer_width = "32")]
-const K: usize = 0x9e3779b9;
+const K: usize = 0x9e37_79b9;
 #[cfg(target_pointer_width = "64")]
-const K: usize = 0x517cc1b727220a95;
+const K: usize = 0x517c_c1b7_2722_0a95;
 
 impl Default for FxHasher {
     #[inline]
-    fn default() -> FxHasher {
-        FxHasher { hash: 0 }
+    fn default() -> Self {
+        Self { hash: 0 }
     }
 }
 
@@ -48,14 +48,16 @@ impl Hasher for FxHasher {
     #[inline]
     fn write(&mut self, mut bytes: &[u8]) {
         #[cfg(target_pointer_width = "32")]
-        let read_usize = |bytes| NativeEndian::read_u32(bytes);
+        #[allow(clippy::cast_possible_truncation)]
+        let read_usize = |bytes| NativeEndian::read_u32(bytes) as usize;
         #[cfg(target_pointer_width = "64")]
-        let read_usize = |bytes| NativeEndian::read_u64(bytes);
+        #[allow(clippy::cast_possible_truncation)]
+        let read_usize = |bytes| NativeEndian::read_u64(bytes) as usize;
 
-        let mut hash = FxHasher { hash: self.hash };
+        let mut hash = Self { hash: self.hash };
         assert!(size_of::<usize>() <= 8);
         while bytes.len() >= size_of::<usize>() {
-            hash.add_to_hash(read_usize(bytes) as usize);
+            hash.add_to_hash(read_usize(bytes));
             bytes = &bytes[size_of::<usize>()..];
         }
         if (size_of::<usize>() > 4) && (bytes.len() >= 4) {
@@ -66,7 +68,7 @@ impl Hasher for FxHasher {
             hash.add_to_hash(NativeEndian::read_u16(bytes) as usize);
             bytes = &bytes[2..];
         }
-        if (size_of::<usize>() > 1) && bytes.len() >= 1 {
+        if (size_of::<usize>() > 1) && !bytes.is_empty() {
             hash.add_to_hash(bytes[0] as usize);
         }
         self.hash = hash.hash;
@@ -89,6 +91,7 @@ impl Hasher for FxHasher {
 
     #[cfg(target_pointer_width = "32")]
     #[inline]
+    #[allow(clippy::cast_possible_truncation)]
     fn write_u64(&mut self, i: u64) {
         self.add_to_hash(i as usize);
         self.add_to_hash((i >> 32) as usize);
@@ -96,6 +99,7 @@ impl Hasher for FxHasher {
 
     #[cfg(target_pointer_width = "64")]
     #[inline]
+    #[allow(clippy::cast_possible_truncation)]
     fn write_u64(&mut self, i: u64) {
         self.add_to_hash(i as usize);
     }