Some targets, like `x86_64-apple-darwin`, always have SSSE3 enabled. Swiss Table has two tiny optimization for SSSE3. This PR implements one of them, see https://github.com/abseil/abseil-cpp/blob/38b704384cd2f17590b3922b97744be0b43622c9/absl/container/internal/raw_hash_set.h#L374.

This PR implements this optimization. Benches before (Haswell):

```
test find_existing     ... bench:       4,160 ns/iter (+/- 1,636)
test find_nonexisting  ... bench:       2,538 ns/iter (+/- 733)
test get_remove_insert ... bench:          29 ns/iter (+/- 22)
test grow_by_insertion ... bench:         164 ns/iter (+/- 28)
test hashmap_as_queue  ... bench:          24 ns/iter (+/- 16)
test new_drop          ... bench:           0 ns/iter (+/- 0)
test new_insert_drop   ... bench:         123 ns/iter (+/- 61)
```

Benches after:

```
test find_existing     ... bench:       3,696 ns/iter (+/- 1,932)
test find_nonexisting  ... bench:       2,566 ns/iter (+/- 725)
test get_remove_insert ... bench:          28 ns/iter (+/- 14)
test grow_by_insertion ... bench:         162 ns/iter (+/- 33)
test hashmap_as_queue  ... bench:          23 ns/iter (+/- 14)
test new_drop          ... bench:           0 ns/iter (+/- 0)
test new_insert_drop   ... bench:         126 ns/iter (+/- 71)
```

The different code gen can be analyzed here: https://rust.godbolt.org/z/iCIKte

I've piped it through LLVM-MCA and Intel's IACA tools, and they show for Skylake:

* Before (SSE2):
    * IACA: cycles/it 1.51, uOps/it 9
    * LLVM-MCA: cycles 804 / 1000 it, uOps/cyc 1.12 (https://rust.godbolt.org/z/20S8c4)

* After (SSSE3):
    * IACA: cycles/it 1.49, uOps/it 9
    * LLVM-MCA SSSE3: cycles 161 / 900 it, uOps/cyc 5.59 (https://rust.godbolt.org/z/0NcwpB)

The model that LLVM-MCA uses looks completely wrong (it predicts the SSSE3 version to be 5x faster), but the model IACA uses looks more realistic, predicting a smaller advantage for the SSSE3 version, which is close to what I've observed in practice. Somebody with Skylake should try to reproduce. 

Basically IACA predicts that both take 9 uOps, but the SSSE3 version puts less pressure on the Ports 1-3:

### SSE2 IACA

```
Intel(R) Architecture Code Analyzer Version -  v3.0-28-g1ba2cbb build date: 2017-10-30;16:57:45
Analyzed File -  sse2
Binary Format - 64Bit
Architecture  -  HSW
Analysis Type - Throughput

Throughput Analysis Report
--------------------------
Block Throughput: 1.51 Cycles       Throughput Bottleneck: FrontEnd
Loop Count:  50
Port Binding In Cycles Per Iteration:
--------------------------------------------------------------------------------------------------
|  Port  |   0   -  DV   |   1   |   2   -  D    |   3   -  D    |   4   |   5   |   6   |   7   |
--------------------------------------------------------------------------------------------------
| Cycles |  0.0     0.0  |  1.0  |  1.0     1.0  |  1.0     1.0  |  1.0  |  1.0  |  1.0  |  1.0  |
--------------------------------------------------------------------------------------------------

DV - Divider pipe (on port 0)
D - Data fetch pipe (on ports 2 and 3)
F - Macro Fusion with the previous instruction occurred
* - instruction micro-ops not bound to a port
^ - Micro Fusion occurred
# - ESP Tracking sync uop was issued
@ - SSE instruction followed an AVX256/AVX512 instruction, dozens of cycles penalty is expected
X - instruction not supported, was not accounted in Analysis

| Num Of   |                    Ports pressure in cycles                         |      |
|  Uops    |  0  - DV    |  1   |  2  -  D    |  3  -  D    |  4   |  5   |  6   |  7   |
-----------------------------------------------------------------------------------------
|   1*     |             |      |             |             |      |      |      |      | pxor xmm0, xmm0
|   2^     |             | 1.0  | 1.0     1.0 |             |      |      |      |      | pcmpgtb xmm0, xmmword ptr [rip-0x8]
|   1*     |             |      |             |             |      |      |      |      | mov rax, rdi
|   2^     |             |      |             | 1.0     1.0 |      | 1.0  |      |      | por xmm0, xmmword ptr [rip+0x3ef8]
|   2^     |             |      |             |             | 1.0  |      |      | 1.0  | movdqa xmmword ptr [rdi], xmm0
|   1      |             |      |             |             |      |      | 1.0  |      | jmp 0xffffffffffffffe5
Total Num Of Uops: 9
```

### SSSE3 IACA

```
Intel(R) Architecture Code Analyzer Version -  v3.0-28-g1ba2cbb build date: 2017-10-30;16:57:45
Analyzed File -  ssse3
Binary Format - 64Bit
Architecture  -  SKL
Analysis Type - Throughput

Throughput Analysis Report
--------------------------
Block Throughput: 1.49 Cycles       Throughput Bottleneck: FrontEnd
Loop Count:  50
Port Binding In Cycles Per Iteration:
--------------------------------------------------------------------------------------------------
|  Port  |   0   -  DV   |   1   |   2   -  D    |   3   -  D    |   4   |   5   |   6   |   7   |
--------------------------------------------------------------------------------------------------
| Cycles |  0.7     0.0  |  0.7  |  1.5     1.5  |  1.5     1.5  |  1.0  |  1.0  |  0.7  |  1.0  |
--------------------------------------------------------------------------------------------------

DV - Divider pipe (on port 0)
D - Data fetch pipe (on ports 2 and 3)
F - Macro Fusion with the previous instruction occurred
* - instruction micro-ops not bound to a port
^ - Micro Fusion occurred
# - ESP Tracking sync uop was issued
@ - SSE instruction followed an AVX256/AVX512 instruction, dozens of cycles penalty is expected
X - instruction not supported, was not accounted in Analysis

| Num Of   |                    Ports pressure in cycles                         |      |
|  Uops    |  0  - DV    |  1   |  2  -  D    |  3  -  D    |  4   |  5   |  6   |  7   |
-----------------------------------------------------------------------------------------
|   1      |             |      | 0.5     0.5 | 0.5     0.5 |      |      |      |      | movdqa xmm0, xmmword ptr [rip+0x3ef6]
|   2^     |             |      | 0.5     0.5 | 0.5     0.5 |      | 1.0  |      |      | pshufb xmm0, xmmword ptr [rsi]
|   1*     |             |      |             |             |      |      |      |      | mov rax, rdi
|   2^     | 0.3         | 0.7  | 0.5     0.5 | 0.5     0.5 |      |      |      |      | por xmm0, xmmword ptr [rip+0x3ef8]
|   2^     |             |      |             |             | 1.0  |      |      | 1.0  | movdqa xmmword ptr [rdi], xmm0
|   1      | 0.3         |      |             |             |      |      | 0.7  |      | jmp 0xffffffffffffffe4
Total Num Of Uops: 9
```