@Amanieu 2019-03-08
`convert_special_to_empty_and_full_to_deleted` is only used when rehashing, so I don't see how this change can improve performance in lookups like your benchmark results show.

The resulting code has the same number of instructions, and IACA shows that there is no significant improvement in cycle count.

Also, are you sure that your implementation is correct? Keep in mind that hashbrown uses a different encoding for the control bytes than SwissTable. My encoding is simpler, which allows the SSE2 code to be just as fast as SwissTable's SSSE3 code.