@gnzlbg 2019-03-08
>  IACA shows that there is no significant improvement in cycle count.

Oops, I posted the results of SSSE3 twice, should be fixed now. 

> so I don't see how this change can improve performance in lookups like your benchmark results show.

No idea. I can reliably reproduce the results. 

> Also, are you sure that your implementation is correct? Keep in mind that hashbrown uses a different encoding for the control bytes than SwissTable.

I assumed the encoding to be similar, but noticed some differences (different masks being used).