@Amanieu 2019-03-08
> I assumed the encoding to be similar, but noticed some differences (different masks being used).

Swisstable uses the following definitions:
```
EMPTY = 0x80
DELETED = 0xFE
SENTINEL = 0xFF
```
hashbrown uses the following definitions:
```
EMPTY = 0xFF
DELETED = 0x80
(we don't use SENTINEL)
```

> No idea. I can reliably reproduce the results.

I've seen this before, the performance difference is due to branch target alignment. If branch targets are aligned to (I think) 16 bytes then the CPU can decode the full 16 bytes of instructions in a single cycle. However if the branch target is misaligned then fewer instruction bytes will be decoded (up to the 16 byte alignment boundary).