@gnzlbg 2019-03-08
> If branch targets are aligned to (I think) 16 bytes then the CPU can decode the full 16 bytes of instructions in a single cycle. However if the branch target is misaligned then fewer instruction bytes will be decoded (up to the 16 byte alignment boundary).

That might explain it.